+++ link/src/main/java/com/stripe/android/link/theme/Shape.kt
@@ -1,11 +1,11 @@
 package com.stripe.android.link.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Shapes
 import androidx.compose.ui.unit.dp
 
-internal val Shapes = Shapes(
-    small = RoundedCornerShape(4.dp),
-    medium = RoundedCornerShape(12.dp),
-    large = RoundedCornerShape(14.dp)
-)
+internal object LinkShapes {
+    val extraSmall = RoundedCornerShape(4.dp)
+    val small = RoundedCornerShape(8.dp)
+    val medium = RoundedCornerShape(12.dp)
+    val large = RoundedCornerShape(14.dp)
+}
